+++ libs/ui/material-autocomplete/src/index.ts
@@ -1,0 +1,2 @@
+export * from './lib/paginated-autocomplete/paginated-autocomplete.component';
+export * from './lib/models/display-fn';