+++ libs/ui/material-autocomplete/src/lib/models/dropdown-data.model.ts
@@ -1,12 +1,12 @@
 /**
- * The data model for the dropdown.
+ * The data model for the autocomplete.
  * For searching the options of the input we really just need the label.
  * This is also used for display.
- * The Data you want to use with the dropdown component just needs to define a label as well.
+ * The Data you want to use with the autocomplete component just needs to define a label as well.
  * Then typescript will allow it even if your data will hold more attributes than the label.
  * This is basically the minimum the component needs.
  */
-export interface DropdownDataModel {
+export interface AutocompleteDataModel {
   /**
    * Label
    */