+++ libs/ui/material-dropdown/src/index.ts
@@ -1,2 +1,0 @@
-export * from './lib/paginated-dropdown/paginated-dropdown.component';
-export * from './lib/models/display-fn';