+++ libs/ui/material-dropdown/src/lib/models/dropdown-type.ts
@@ -1,5 +1,0 @@
-export const DropdownVariants = {
-  COMPACT: 'Compact',
-  CARD: 'Card',
-  DROPDOWN_WITH_CHECKBOXES: 'DropdownWithCheckboxes',
-};